step 3 organized vulkan_data
