+++ src/teapot/src/vulkan/mod.rs
@@ -1,3 +1,7 @@
 mod vulkan_data;
+mod vulkan_data_fns;
 
 pub use vulkan_data::*;
+pub use vulkan_data_fns::*;
+
+
